+++ src/Math/Vector2.cpp
@@ -111,7 +111,7 @@ Vector2& Vector2::operator/=(double s)
 	return *this;
 }
 
-double Vector2::get(uint index) const
+double Vector2::get(uint64_t index) const
 {
 	switch (index)
 	{
@@ -121,7 +121,7 @@ double Vector2::get(uint index) const
 	}
 }
 
-void Vector2::set(uint index, double value)
+void Vector2::set(uint64_t index, double value)
 {
 	switch (index)
 	{
