+++ src/Math/Vector2.h
@@ -34,8 +34,8 @@ namespace Raycer
 		Vector2& operator/=(const Vector2& v);
 		Vector2& operator/=(double s);
 
-		double get(uint index) const;
-		void set(uint index, double value);
+		double get(uint64_t index) const;
+		void set(uint64_t index, double value);
 		double length() const;
 		double lengthSquared() const;
 		void normalize();
